backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation
pydantic==2.7.1
pydantic-settings==2.2.1

# HTTP client
httpx==0.27.0

# Environment
python-dotenv==1.0.1

# Parsing
tree-sitter==0.21.3

# Vector storage (Phase 2)
chromadb==0.5.0

# Dev / testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0